--- src/main/java/br/com/authentication/AuthenticationApplication.java
@@ -1,13 +1,0 @@
-package br.com.authentication;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthenticationApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthenticationApplication.class, args);
-	}
-
-}
