--- src/main/java/br/com/authentication/domain/Role.java
@@ -1,32 +1,0 @@
-package br.com.authentication.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.util.UUID;
-
-@Entity
-@Table(name = "role")
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class Role {
-
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
-    private UUID id;
-
-    private String name;
-
-    public Role(UUID id) {
-        this.id = id;
-    }
-
-}
