--- src/main/java/br/com/authentication/domain/UserAccount.java
@@ -1,54 +1,0 @@
-package br.com.authentication.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.UUID;
-
-@Entity
-@Table(name = "user_account")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserAccount {
-
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
-    private UUID id;
-
-    @Column(name = "first_name", nullable = false)
-    private String firstName;
-
-    @Column(name = "last_name", nullable = false)
-    private String lastName;
-
-    @Column(name = "username", nullable = false)
-    private String username;
-
-    @Column(name = "email", nullable = false, unique = true)
-    private String email;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Column(name = "birthdate")
-    private LocalDate birthdate;
-
-    @Column(name = "phone_1")
-    private String phone1;
-
-    @Column(name = "phone_2")
-    private String phone2;
-
-    @ManyToMany
-    private List<Role> roles;
-
-}
