--- src/main/java/br/com/authentication/dto/AuthenticateUserDto.java
@@ -1,22 +1,0 @@
-package br.com.authentication.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticateUserDto {
-
-    @NotBlank(message = "E-mail é obrigatório")
-    @Email(message = "E-mail inválido")
-    private String email;
-
-    @NotBlank(message = "Senha é obrigatória")
-    private String password;
-
-}
