--- src/main/java/br/com/authentication/dto/UserRoleCreatedDto.java
@@ -1,13 +1,0 @@
-package br.com.authentication.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class UserRoleCreatedDto {
-
-    private String username;
-    private List<String> roles;
-
-}
