--- src/main/java/br/com/authentication/exception/BadRequestExceptionDetails.java
@@ -1,9 +1,0 @@
-package br.com.authentication.exception;
-
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-public class BadRequestExceptionDetails extends ExceptionDetails {
-}
