--- src/main/java/br/com/authentication/exception/ExceptionDetails.java
@@ -1,16 +1,0 @@
-package br.com.authentication.exception;
-
-import lombok.Data;
-import lombok.experimental.SuperBuilder;
-
-import java.time.LocalDateTime;
-
-@Data
-@SuperBuilder
-public class ExceptionDetails {
-    protected String title;
-    protected int status;
-    protected String details;
-    protected String developerMessage;
-    protected LocalDateTime timestamp;
-}
