--- src/main/java/br/com/authentication/exception/ValidationExceptionDetails.java
@@ -1,11 +1,0 @@
-package br.com.authentication.exception;
-
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-public class ValidationExceptionDetails extends ExceptionDetails {
-    private final String fields;
-    private final String fieldsMessage;
-}
