--- src/main/java/br/com/authentication/exception/handler/RestExceptionHandler.java
@@ -1,72 +1,0 @@
-package br.com.authentication.exception.handler;
-
-import br.com.authentication.exception.BadRequestException;
-import br.com.authentication.exception.BadRequestExceptionDetails;
-import br.com.authentication.exception.ExceptionDetails;
-import br.com.authentication.exception.ValidationExceptionDetails;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@ControllerAdvice
-public class RestExceptionHandler extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler(BadRequestException.class)
-    public ResponseEntity<BadRequestExceptionDetails> handlerBadRequestException(BadRequestException bre) {
-        return new ResponseEntity<>(
-                BadRequestExceptionDetails.builder()
-                        .timestamp(LocalDateTime.now())
-                        .status(HttpStatus.BAD_REQUEST.value())
-                        .title("Bad Request Exception, check the details")
-                        .details(bre.getMessage())
-                        .developerMessage(bre.getClass().getName())
-                        .build(), HttpStatus.BAD_REQUEST);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException exception,
-                                                                  HttpHeaders headers, HttpStatus status, WebRequest request) {
-
-        List<FieldError> fieldErrors = exception.getBindingResult().getFieldErrors();
-        String fields = fieldErrors.stream().map(FieldError::getField).collect(Collectors.joining(", "));
-        String fieldsMessage = fieldErrors.stream().map(FieldError::getDefaultMessage).collect(Collectors.joining(", "));
-
-        return new ResponseEntity<>(
-                ValidationExceptionDetails.builder()
-                        .timestamp(LocalDateTime.now())
-                        .status(HttpStatus.BAD_REQUEST.value())
-                        .title("Bad Request Exception, invalid fields")
-                        .details("Check field(s)")
-                        .developerMessage(exception.getClass().getName())
-                        .fields(fields)
-                        .fieldsMessage(fieldsMessage)
-                        .build(), HttpStatus.BAD_REQUEST);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleExceptionInternal(Exception ex, @Nullable Object body, HttpHeaders headers,
-                                                             HttpStatus status, WebRequest request) {
-        ExceptionDetails exceptionDetails = ExceptionDetails.builder()
-                .timestamp(LocalDateTime.now())
-                .status(status.value())
-                .title(ex.getCause().getMessage())
-                .details(ex.getMessage())
-                .developerMessage(ex.getClass().getName())
-                .build();
-
-        return new ResponseEntity<>(exceptionDetails, headers, status);
-    }
-
-
-}
