--- src/main/java/br/com/authentication/repository/RoleRepository.java
@@ -1,12 +1,0 @@
-package br.com.authentication.repository;
-
-import br.com.authentication.domain.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, UUID> {
-
-}
