--- src/main/java/br/com/authentication/repository/UserRepository.java
@@ -1,18 +1,0 @@
-package br.com.authentication.repository;
-
-import br.com.authentication.domain.UserAccount;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserAccount, UUID> {
-
-    Optional<UserAccount> findByEmail(String email);
-
-    Optional<UserAccount> findByUsername(String username);
-
-    boolean existsByUsername(String username);
-}
