--- src/main/java/br/com/authentication/validator/UserValidator.java
@@ -1,8 +1,0 @@
-package br.com.authentication.validator;
-
-public interface UserValidator {
-
-    void validateSamePassword(String password1, String password2);
-
-    void validateUsernameAlreadyExists(String username);
-}
