+++ src/main/java/br/com/planner/authentication/AuthenticationApplication.java
@@ -1,0 +1,13 @@
+package br.com.planner.authentication;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthenticationApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthenticationApplication.class, args);
+	}
+
+}
