+++ src/main/java/br/com/planner/authentication/dto/AuthenticateUserDto.java
@@ -1,0 +1,20 @@
+package br.com.planner.authentication.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticateUserDto {
+
+    @NotBlank(message = "Username é obrigatório")
+    private String username;
+
+    @NotBlank(message = "Senha é obrigatória")
+    private String password;
+
+}
