+++ src/main/java/br/com/planner/authentication/dto/UserRoleCreatedDto.java
@@ -1,0 +1,13 @@
+package br.com.planner.authentication.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class UserRoleCreatedDto {
+
+    private String username;
+    private List<String> roles;
+
+}
