+++ src/main/java/br/com/planner/authentication/exception/BadRequestException.java
@@ -1,0 +1,13 @@
+package br.com.planner.authentication.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class BadRequestException extends RuntimeException {
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+
+}
