+++ src/main/java/br/com/planner/authentication/exception/BadRequestExceptionDetails.java
@@ -1,0 +1,9 @@
+package br.com.planner.authentication.exception;
+
+import lombok.Getter;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+public class BadRequestExceptionDetails extends ExceptionDetails {
+}
