+++ src/main/java/br/com/planner/authentication/exception/ExceptionDetails.java
@@ -1,0 +1,16 @@
+package br.com.planner.authentication.exception;
+
+import lombok.Data;
+import lombok.experimental.SuperBuilder;
+
+import java.time.LocalDateTime;
+
+@Data
+@SuperBuilder
+public class ExceptionDetails {
+    protected String title;
+    protected int status;
+    protected String details;
+    protected String developerMessage;
+    protected LocalDateTime timestamp;
+}
