+++ src/main/java/br/com/planner/authentication/exception/ValidationExceptionDetails.java
@@ -1,0 +1,11 @@
+package br.com.planner.authentication.exception;
+
+import lombok.Getter;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+public class ValidationExceptionDetails extends ExceptionDetails {
+    private final String fields;
+    private final String fieldsMessage;
+}
