+++ src/main/java/br/com/planner/authentication/repository/RoleRepository.java
@@ -1,0 +1,12 @@
+package br.com.planner.authentication.repository;
+
+import br.com.planner.authentication.domain.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, UUID> {
+
+}
