+++ src/main/java/br/com/planner/authentication/service/EncoderService.java
@@ -1,0 +1,11 @@
+package br.com.planner.authentication.service;
+
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+public interface EncoderService {
+
+    BCryptPasswordEncoder getEnconder();
+
+    String encode(String password);
+
+}
