+++ src/main/java/br/com/planner/authentication/service/UserRoleService.java
@@ -1,0 +1,14 @@
+package br.com.planner.authentication.service;
+
+import br.com.planner.authentication.domain.Role;
+import br.com.planner.authentication.dto.CreateUserRoleDto;
+import br.com.planner.authentication.dto.UserRoleCreatedDto;
+
+import java.util.List;
+
+public interface UserRoleService {
+
+    UserRoleCreatedDto saveUserRole(CreateUserRoleDto dto);
+
+    List<Role> findAll();
+}
