+++ src/main/java/br/com/planner/authentication/service/UserService.java
@@ -1,0 +1,14 @@
+package br.com.planner.authentication.service;
+
+import br.com.planner.authentication.domain.UserAccount;
+import br.com.planner.authentication.dto.RegisterUserDto;
+
+import java.util.Optional;
+
+public interface UserService {
+
+    void registerUserAccount(RegisterUserDto userAccount);
+
+    Optional<UserAccount> getByUsername(String username);
+
+}
