--- src/test/java/br/com/authentication/AuthenticationApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.authentication;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthenticationApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
