+++ src/test/java/br/com/planner/authentication/AuthenticationApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.planner.authentication;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuthenticationApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
